jash intern/farmsetu_weather/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
python-dotenv==1.0.1
requests==2.32.3
